Refs #61 - Escape using backslash inside { text }
This fixes half the problem specified in #61.  The doc change
is technically inaccurate - opening brackets don't actually NEED to
be escaped - p{{\}} works as well as p{\{\}}.  But they CAN be escaped,
so I used the most regular version.

The part that is NOT fixed is that a | character, escaped or not, will
prevent the expression from being parsed AT ALL.