Added multi-line * wrap functionality as per #81
This is a proof-of-concept - it needs to do the token-replacement jazz the original
version did, because right now this will fail if the content of the
txt/lines includes unmatched braces in various configurations.

I think the overall plan is sound: split expression into leading and
terminal components, render the text as a list (either a one-element
list of all text or a multiline list, depending on trailing *), and
create a terminal phrase consisting of "terminal-element{TOKEN}" joined
on "+"

To do: render list into a-list of '((hash-of-txt . txt), ...), mapconcat
the hash replacement instead of the text, then run replace over result.
Write tests.  Also, tests appear to be broken on master, at least on my machine.