Implement token-replacement in new wrap * function
It itched at me, so I cleaned it up and rewrote to make it do the token
replacement.  Would be more performant to do the sha1 once, but
realistically, it's not going to be meaningfully slow for the use-case.

This is essentially complete, but needs tests - I'll add them as soon as
I can get testing working on my computer.